routeファイルにexportできない関数があったので切り出す
